fix: edx_health and py_dependency_health markers implemented in same folder
