--- health_checks/dependencies_health/__init__.py
@@ -1,5 +1,0 @@
-"""
-A module that contains all the checks for dependencies dashboard
-"""
-
-__version__ = "1.0.0"
