+++ repo_health/check_requirements.py
@@ -37,7 +37,7 @@ def fixture_req_lines(repo_path):
     },
 )
 @pytest.mark.py_dependency_health
-@pytest.mark.repo_health
+@pytest.mark.edx_health
 def check_requires(req_lines, all_results):
     """
     Test to find whether repo requires some key requirements
