+++ repo_health/check_ubuntufiles.py
@@ -312,6 +312,7 @@ def fixture_ubuntu_content(repo_path, git_origin_url):
         "apt_get_packages": "content name published on ubuntu.",
         "yml_files": "content name published on ubuntu.",
     })
+@pytest.mark.edx_health
 def check_ubuntu_content(content, all_results):
     """
     Adding data into results.
