+++ repo_health_dashboard/dependencies_configuration.yaml
